+++ CHANGELOG.md
@@ -11,12 +11,19 @@ and this project adheres to
 ### Changed
 
 - 📦️(front) update react
+- ✨(front) add ui-kit
+- ✨(front) add dark-mode
+
 
 ### Fixed
 
 - 🐛(e2e) fix test-e2e-chromium
 - 🐛(back) fix system prompt compatibility with self-hosted models #200
 - ⚰️(back) remove dead code and unused files
+
+### Removed
+
+- 🔥(chat) remove thinking part from frontend #227
 
 ## [0.0.10] - 2025-12-15